% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix-pin.R
\name{connection_pin_read}
\alias{connection_pin_read}
\title{Read a pinned dbMatrix object from a board}
\usage{
connection_pin_read(board, name, version = NULL)
}
\arguments{
\item{board}{A pins board object}

\item{name}{The name of the pin}

\item{version}{The version of the pin to get (optional)}
}
\value{
A dbMatrix object (dbSparseMatrix or dbDenseMatrix)
}
\description{
Read a pinned dbMatrix object from a board
}
