% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dbData-class}
\alias{dbData-class}
\title{Virtual Base Class for Database-Backed Data Objects}
\description{
\code{dbData} is a virtual base class for database-backed objects in the dbverse ecosystem.
This class provides a common interface for objects that are backed by database tables.
It is extended by concrete classes like dbMatrix and dbSpatial that provide specific
implementations.
}
\section{Slots}{

\describe{
\item{\code{value}}{tbl_duckdb_connection that represents the data in the database}

\item{\code{name}}{name of table within database that contains the data}
}}

\section{Extensibility}{

When creating a new type of database-backed object, extend this class to ensure
compatibility with the dbverse ecosystem.
}

