% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dbLoad}
\alias{dbLoad}
\title{Load a dbverse object from the database}
\usage{
dbLoad(conn, name, class, ...)
}
\arguments{
\item{conn}{A DBI database connection}

\item{name}{Character name of the table/view in the database}

\item{class}{Character class name of the object to load (e.g., "dbMatrix", "dbSpatial")}

\item{...}{Additional arguments passed to specific methods}
}
\value{
A dbverse object of the specified class
}
\description{
Generic function to reconstruct dbverse objects (dbMatrix, dbSpatial, etc.)
from tables stored in a database connection.
}
