% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/connection-methods.R
\name{dbReconnect}
\alias{dbReconnect}
\alias{dbReconnect,dbData-method}
\alias{dbReconnect,DBIConnection-method}
\alias{dbReconnect,connConnection-method}
\title{Reconnect database connection for dbData objects}
\usage{
dbReconnect(x)

\S4method{dbReconnect}{dbData}(x)

\S4method{dbReconnect}{DBIConnection}(x)

\S4method{dbReconnect}{connConnection}(x)
}
\arguments{
\item{x}{A dbData object}
}
\value{
The dbData object with an updated connection
}
\description{
Reconnects invalid database connections for dbData objects, updating the
object with the new valid connection.
}
