% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-validation.R
\name{.check_con}
\alias{.check_con}
\title{Validate database connection}
\usage{
.check_con(conn)
}
\arguments{
\item{conn}{A DBI connection object}
}
\description{
Validate database connection
}
\keyword{internal}
