% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-validation.R
\name{.check_tbl}
\alias{.check_tbl}
\title{Validate dplyr tbl object}
\usage{
.check_tbl(tbl)
}
\arguments{
\item{tbl}{A tbl_duckdb_connection object}
}
\description{
Validate dplyr tbl object
}
\keyword{internal}
