% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix-pin.R
\name{read_pin_conn.conn_matrix_table}
\alias{read_pin_conn.conn_matrix_table}
\title{Read a pinned dbMatrix from pins board}
\usage{
\method{read_pin_conn}{conn_matrix_table}(x)
}
\arguments{
\item{x}{A pinned conn_matrix_table object}
}
\value{
A dbMatrix object (dbSparseMatrix or dbDenseMatrix)
}
\description{
S3 method for reading dbMatrix objects from a pins board.
Reconstructs the full dbMatrix object.
}
