% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial-pin.R
\name{read_pin_conn.conn_spatial_table}
\alias{read_pin_conn.conn_spatial_table}
\title{Read a pinned dbSpatial object from a pins board}
\usage{
\method{read_pin_conn}{conn_spatial_table}(x)
}
\arguments{
\item{x}{A pinned conn_spatial_table object}
}
\value{
A dbSpatial object
}
\description{
Read a pinned dbSpatial object from a pins board
}
