% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix-pin.R
\name{read_pin_conn}
\alias{read_pin_conn}
\title{Read Pin Connection Generic}
\usage{
read_pin_conn(x)
}
\arguments{
\item{x}{A pinned connection object}
}
\description{
Generic function for reading connection objects from a pins board
}
