% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_view.R
\name{to_view,ANY-method}
\alias{to_view,ANY-method}
\title{Convert lazy table to named VIEW}
\usage{
\S4method{to_view}{ANY}(x, name, temporary = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{x}{\code{tbl_sql}
\strong{Required}. \code{tbl_sql} object to convert to a VIEW.}

\item{name}{\code{character}
\strong{Required}. Name to assign VIEW within database. Auto-generated if none
provided with prefix \code{"tmp_view_"}}

\item{temporary}{\code{logical}
If \code{TRUE} (default), the VIEW will not be saved in the database}

\item{overwrite}{\code{logical}
If \code{TRUE} (default), the VIEW will overwrite an existing VIEW of the same name}

\item{...}{Additional arguments passed to \code{\link[DBI:dbExecute]{DBI::dbExecute()}}}
}
\description{
Convert lazy table to named VIEW
}
\concept{dbData}
