% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{to_view}
\alias{to_view}
\title{Convert lazy table to named view}
\usage{
to_view(x, name, temporary = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{x}{A lazy table object (tbl_duckdb_connection)}

\item{name}{Character name to assign view within database}

\item{temporary}{Logical, if TRUE (default), the view will be deleted after the session ends}

\item{overwrite}{Logical, if TRUE (default), the view will overwrite an existing view}

\item{...}{Additional arguments passed to DBI::dbExecute}
}
\description{
Convert lazy table to named view
}
