% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-validation.R
\name{unique_table_name}
\alias{unique_table_name}
\title{Generate unique table names}
\usage{
unique_table_name(prefix = "db")
}
\arguments{
\item{prefix}{Character prefix for table name}
}
\value{
Character, unique table name
}
\description{
Generate unique table names
}
\keyword{internal}
