% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSpatial-pin.R
\name{write_pin_conn.dbSpatial}
\alias{write_pin_conn.dbSpatial}
\title{Write a dbSpatial object to a pins board}
\usage{
\method{write_pin_conn}{dbSpatial}(x, board, name, ...)
}
\arguments{
\item{x}{A dbSpatial object}

\item{board}{A pins board object}

\item{name}{Name for the pin}

\item{...}{Additional arguments passed to pins::pin_write}
}
\value{
Invisibly returns the input object
}
\description{
Write a dbSpatial object to a pins board
}
