% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbMatrix-pin.R
\name{write_pin_conn}
\alias{write_pin_conn}
\title{Write Pin Connection Generic}
\usage{
write_pin_conn(x, board, ...)
}
\arguments{
\item{x}{Object to write}

\item{board}{A pins board object}

\item{...}{Additional arguments passed to methods}
}
\description{
Generic function for writing connection objects to a pins board
}
